backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
python-jose[cryptography]
passlib[bcrypt]==1.7.4
python-multipart
python-dotenv
httpx
websockets
groq
email-validator
pydantic[email]
bcrypt==4.0.1
python-socketio[async_client]
yt-dlp
python-docx
reportlab
markdown
pydantic
transformers
torch
torchaudio
librosa
soundfile
pydub
PyPDF2
google-generativeai
python-pptx
tiktoken>=0.5.0
google-genai
opencv-python
pillow
numpy
scikit-image
